add style checker, fix style
