+++ .idea/compiler.xml
@@ -10,7 +10,12 @@
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
-      <module name="evvo" target="1.8" />
+      <module name="evvo" target="1.6" />
     </bytecodeTargetLevel>
+  </component>
+  <component name="JavacSettings">
+    <option name="ADDITIONAL_OPTIONS_OVERRIDE">
+      <module name="evvo" options="-Werror" />
+    </option>
   </component>
 </project>
