+++ src/main/scala/com/diatom/agent/deletor.scala
