+++ src/main/scala/com/diatom/agent/strategy.scala
