+++ src/main/scala/com/diatom/island/population/scored.scala
