+++ src/main/scala/com/diatom/package.scala
